Only download Cake if necessary.

+++ build.ps1
@@ -19,14 +19,6 @@ $CakeVersion = "0.30.0"
 $CakeDirPath = Join-Path $PSScriptRoot "cake"
 New-Item -Path $CakeDirPath -Type Directory -ErrorAction SilentlyContinue | Out-Null
 
-# create packages.config
-$PackagesConfigPath = Join-Path $CakeDirPath "packages.config"
-[System.IO.File]::WriteAllLines($PackagesConfigPath, @(
-    "<?xml version=`"1.0`" encoding=`"utf-8`"?>",
-    "<packages>",
-    "`t<package id=`"Cake`" version=`"$CakeVersion`" />",
-    "</packages>"))
-
 # download nuget.exe if not in path and not already downloaded
 $NuGetExe = Get-Command "nuget.exe" -ErrorAction SilentlyContinue
 If ($NuGetExe -ne $null) {
@@ -39,16 +31,27 @@ Else {
     }
 }
 
-# use NuGet to download Cake
-Push-Location $CakeDirPath
-Invoke-Expression "&`"$NuGetExePath`" install -OutputDirectory ."
-If ($LASTEXITCODE -ne 0) {
-    Throw "An error occurred while restoring NuGet tools."
+# download Cake if necessary
+$CakeExePath = Join-Path $CakeDirPath "Cake.$CakeVersion/Cake.exe"
+If (!(Test-Path $CakeExePath)) {
+    # create packages.config
+    $PackagesConfigPath = Join-Path $CakeDirPath "packages.config"
+    [System.IO.File]::WriteAllLines($PackagesConfigPath, @(
+        "<?xml version=`"1.0`" encoding=`"utf-8`"?>",
+        "<packages>",
+        "`t<package id=`"Cake`" version=`"$CakeVersion`" />",
+        "</packages>"))
+
+    # use NuGet to download Cake
+    Push-Location $CakeDirPath
+    Invoke-Expression "&`"$NuGetExePath`" install -OutputDirectory ."
+    If ($LASTEXITCODE -ne 0) {
+        Throw "An error occurred while restoring NuGet tools."
+    }
+    Pop-Location
 }
-Pop-Location
 
 # run Cake with specified arguments
-$CakeExePath = Join-Path $CakeDirPath "Cake.$CakeVersion/Cake.exe"
 $ExtraArgs = ""
 if ($Target) { $ExtraArgs += "--target=$Target" }
 Invoke-Expression "& `"$CakeExePath`" --paths_tools=cake $ExtraArgs $ScriptArgs"
